Save health and player position

+++ KVLT/PlayerWeapon.h
@@ -3,7 +3,7 @@ class PlayerWeapon{
 private:
 	enum _WeaponWeight {LIGHT, MEDIUM, HEAVY};
 	_WeaponWeight currentWW = LIGHT;
-	int _weaponDamage = 0;
+	unsigned short int _weaponDamage = 0;
 	float _weaponSpeed = 0;
 	int GetWeaponWeight();
 public:
